Added tests for the menu storage methods

+++ project/backend/app/repositories/menu_items_repos.py
@@ -15,10 +15,14 @@ def load_menu(restaurant_id: int) -> Dict[Any, Dict[Any, Any]]:
         items = {}
         for row in reader:
             if (int(row["restaurant_id"]) == restaurant_id):
-                items[row["restaurant_id"]] = row
-                items[row["item_id"]].pop("restaurant_id")
-        return items
-    
+                item_id = row["item_id"]
+                items[item_id] = row
+                items[item_id].pop("restaurant_id")
+        if (len(items) > 0):
+            return items
+        else:
+            raise IndexError(f"Error: Unable to find menu items for restaurant id:{restaurant_id}")
+        
 def load_menu_item(restaurant_id: int, item_id: int) -> Dict[Any, Any]:
     if not DATA_PATH.exists():
         raise FileExistsError(
@@ -33,9 +37,23 @@ def load_menu_item(restaurant_id: int, item_id: int) -> Dict[Any, Any]:
 
 def save_menu(restaurant_id: int, items: Dict[Any, Dict[Any, Any]]) -> None:
     fieldNames = ['restaurant_id','item_id','item_name','price','description','image']
+    if not DATA_PATH.exists():
+        raise FileExistsError(
+            "Error: The storage csv does not exist or otherwise cannot be accessed"
+            )
+
+    existing_rows = []
+    with DATA_PATH.open("r", encoding="utf-8", newline='') as f:
+        reader = csv.DictReader(f)
+        existing_rows = list(reader)
+    existing_rows = [row for row in existing_rows if row.get("restaurant_id") != str(restaurant_id)]
+
     with DATA_PATH.open("w", encoding="utf-8", newline='') as f:
         writer = csv.DictWriter(f, fieldnames=fieldNames)
         writer.writeheader()
-        for row in items:
-            row['restaurant_id'] = restaurant_id
-            writer.writerow(row)
+        new_rows = []
+        for i in range(1,len(items)+1):
+            items[str(i)]["restaurant_id"] = str(restaurant_id)
+            new_rows.append(items[str(i)])
+        writer.writerows(existing_rows + new_rows)
+
